GH-4 Shifting around test code, mock structs, and benchmarks. Corrected a couple issues and a race condition
