+++ .travis.yml
@@ -1,4 +1,9 @@
 language: go
+
+go:
+  - 1.13.x
+  - master
+
 env:
   - GO111MODULE=on
